+++ src/main/java/org/springeel/oneclickrecipe/domain/order/controller/OrderController.java
@@ -5,7 +5,7 @@ import org.springeel.oneclickrecipe.domain.order.dto.controller.OrderCreateContr
 import org.springeel.oneclickrecipe.domain.order.dto.service.OrderCreateResponseDto;
 import org.springeel.oneclickrecipe.domain.order.entity.Order;
 import org.springeel.oneclickrecipe.domain.order.mapper.OrderMapper;
-import org.springeel.oneclickrecipe.domain.order.service.OrderService;
+import org.springeel.oneclickrecipe.domain.order.service.impl.OrderServiceImpl;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -19,7 +19,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class OrderController {
 
     private final OrderMapper orderMapper;
-    private final OrderService orderService;
+    private final OrderServiceImpl orderService;
 
     @PostMapping("/orders")
     public ResponseEntity<OrderCreateResponseDto> createOrder(
